Move deterministic Workflow Session observations into canonical crate
